Fix the vnf test
